Crawl/requirements.txt
requests 
beautifulsoup4 
tqdm
cloudinary
psycopg2-binary
dotenv
unidecode
tenacity
aiohttp